api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
boto3==1.34.162
psycopg2-binary==2.9.9
python-multipart==0.0.9
